Migrate cortland (rotated-SOVC PDF, positional 4-block Total slice)
